+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@svgr/webpack": "^8.1.0",
+    "gsap": "^3.12.5",
     "next": "14.2.15",
     "react": "^18",
     "react-dom": "^18"
