+++ src/app/shared/constants.ts
@@ -1,0 +1,20 @@
+export const constants = {
+    heading1: "Creating Stunning Designs",
+    heading2: "Designing Dreams into Digital Realities",
+    description: "A passionate UX designer dedicated to crafting immersive digital experiences that blend creativity with functionality. With a keen eye for detail and a commitment to user-centric design",
+    years: "2 Years",
+    ofExperience: "of Experience",
+    shapeExperience: "Shaping Experiences Through Minimalist Design",
+    worksOfArt: "Works of Art that are Simply a Treat to Use",
+    standardAutomobile: "Standard Automobile App Revamp",
+    standardAutomobileContent: "Completely revamped the UX design of Standard Automobile app for UAE. Improved the over all user experience. Rectified more then 235 heuristic issues within the app. Redesigned to a stunning design focusing on the current trends.",
+    smartSchedule: "Smart scheduling web application",
+    smartScheduleContent: "Web application which helps people grow their connection with a discover and calendar function.",
+    luxuryAutomobile: "Luxury Automobile App Revamp",
+    luxuryContent: "Completely revamped the UX design of Luxury Automobile app for UAE",
+    masterpieces: "Let’s create some masterpieces together",
+    portOperation: "Port operation software",
+    portOperationContent: "Software that streamlines marine logistics and enhances overall efficiency in port management.",
+    usedCar: "Used Cars Dealership Website",
+    usedCarContent: "Website which focus on ease and efficiency of buying and selling used cars."
+};
