+++ src/app/shared/resources.ts
@@ -1,0 +1,3 @@
+export const resources = {
+    homeGif: "https://i.giphy.com/media/v1.Y2lkPTc5MGI3NjExc29iMjVkYzJmMXdxcWY4NXdzM2c2ZXVjdjFnamt3Z2Q2cmk0dW0weiZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/GrWIUslw4cWqBtiAIJ/giphy-downsized-large.gif",
+};
